requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
entrypoints==0.3
flake8==3.7.7
flake8-mypy==17.8.0
Flask==1.0.2
google-api-core==1.8.1
google-auth==1.6.3
google-cloud-core==0.29.1
google-cloud-storage==1.14.0
google-resumable-media==0.3.2
googleapis-common-protos==1.5.8
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==6.0.0
mypy==0.670
mypy-extensions==0.4.1
numpy==1.16.2
pandas==0.24.2
pluggy==0.9.0
protobuf==3.7.0
py==1.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycodestyle==2.5.0
pyflakes==2.1.1
pytest==4.3.1
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
rsa==4.0
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
sklearn==0.0
typed-ast==1.3.1
urllib3==1.24.1
Werkzeug==0.14.1
